Refactor : window 새로고침 함수 삭제 & 불필요한 컴포넌트 파일 삭제
